+++ docs/.vuepress/components/endpoints/data/attachment.js
@@ -4,11 +4,11 @@ export default [
     type: 'string',
     description: 'The attachment ID',
   },
-  {
-    field: 'name',
-    type: 'string',
-    description: 'The attachment resource name',
-  },
+  // {
+  //   field: 'name',
+  //   type: 'string',
+  //   description: 'The attachment resource name',
+  // },
   {
     field: 'alt',
     type: 'string',
@@ -19,24 +19,24 @@ export default [
     type: 'string',
     description: 'The public url of the attachment',
   },
-  {
-    field: 'disk',
-    type: 'string',
-    description: 'The file stroage disk',
-  },
-  {
-    field: 'size',
-    type: 'number',
-    description: 'The file size, in bytes',
-  },
-  {
-    field: 'pending',
-    type: 'boolean',
-    description: 'The attachment persistence state',
-  },
-  {
-    field: 'location',
-    type: 'string',
-    description: 'The attachment location. This value is reflective of the storage absolute path',
-  },
+  // {
+  //   field: 'disk',
+  //   type: 'string',
+  //   description: 'The file stroage disk',
+  // },
+  // {
+  //   field: 'size',
+  //   type: 'number',
+  //   description: 'The file size, in bytes',
+  // },
+  // {
+  //   field: 'pending',
+  //   type: 'boolean',
+  //   description: 'The attachment persistence state',
+  // },
+  // {
+  //   field: 'location',
+  //   type: 'string',
+  //   description: 'The attachment location. This value is reflective of the storage absolute path',
+  // },
 ]