+++ docs/.vuepress/components/endpoints/data/share.js
@@ -63,6 +63,11 @@ export default {
   ],
   response: [
     {
+      field: 'id',
+      type: 'string',
+      description: 'The shared content resource ID',
+    },
+    {
       field: 'url',
       type: 'string',
       description: 'A url to share with external',