+++ CHANGELOG.md
@@ -7,3 +7,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - fix: `/downloadable-link/{file}` endpoint should receive `POST` requests
 - feat: add `share` REST endpoints
 - fix: content sharing defaults not to override
+
+## 1.0.2
+- fix: add `description` pipe to `MOIREI\MediaLibrary\Upload`
+- fix: attachment API controller to return `id` and `alt` after store